+++ components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -346,11 +346,6 @@ class GeckoEngineTest {
             CookiePolicy.ACCEPT_FIRST_PARTY_AND_ISOLATE_OTHERS.id,
         )
 
-        assertEquals(contentBlockingSettings.cookieBannerMode, EngineSession.CookieBannerHandlingMode.DISABLED.mode)
-        assertEquals(contentBlockingSettings.cookieBannerModePrivateBrowsing, EngineSession.CookieBannerHandlingMode.DISABLED.mode)
-        assertEquals(contentBlockingSettings.cookieBannerDetectOnlyMode, engine.settings.cookieBannerHandlingDetectOnlyMode)
-        assertEquals(contentBlockingSettings.cookieBannerGlobalRulesEnabled, engine.settings.cookieBannerHandlingGlobalRules)
-        assertEquals(contentBlockingSettings.cookieBannerGlobalRulesSubFramesEnabled, engine.settings.cookieBannerHandlingGlobalRulesSubFrames)
         assertEquals(contentBlockingSettings.queryParameterStrippingEnabled, engine.settings.queryParameterStripping)
         assertEquals(contentBlockingSettings.queryParameterStrippingPrivateBrowsingEnabled, engine.settings.queryParameterStrippingPrivateBrowsing)
         assertEquals(contentBlockingSettings.queryParameterStrippingAllowList[0], engine.settings.queryParameterStrippingAllowList)
@@ -713,108 +708,6 @@ class GeckoEngineTest {
     }
 
     @Test
-    fun `setCookieBannerMode is only invoked when the value is changed`() {
-        val mockRuntime = mock<GeckoRuntime>()
-        val settings = spy(ContentBlocking.Settings.Builder().build())
-        whenever(mockRuntime.settings).thenReturn(mock())
-        whenever(mockRuntime.settings.contentBlocking).thenReturn(settings)
-
-        val engine = GeckoEngine(testContext, runtime = mockRuntime)
-        val policy = EngineSession.CookieBannerHandlingMode.REJECT_ALL
-
-        engine.settings.cookieBannerHandlingMode = policy
-
-        verify(mockRuntime.settings.contentBlocking).setCookieBannerMode(policy.mode)
-
-        reset(settings)
-
-        engine.settings.cookieBannerHandlingMode = policy
-
-        verify(mockRuntime.settings.contentBlocking, never()).setCookieBannerMode(policy.mode)
-    }
-
-    @Test
-    fun `setCookieBannerModePrivateBrowsing is only invoked when the value is changed`() {
-        val mockRuntime = mock<GeckoRuntime>()
-        val settings = spy(ContentBlocking.Settings.Builder().build())
-        whenever(mockRuntime.settings).thenReturn(mock())
-        whenever(mockRuntime.settings.contentBlocking).thenReturn(settings)
-
-        val engine = GeckoEngine(testContext, runtime = mockRuntime)
-        val policy = EngineSession.CookieBannerHandlingMode.REJECT_OR_ACCEPT_ALL
-
-        engine.settings.cookieBannerHandlingModePrivateBrowsing = policy
-
-        verify(mockRuntime.settings.contentBlocking).setCookieBannerModePrivateBrowsing(policy.mode)
-
-        reset(settings)
-
-        engine.settings.cookieBannerHandlingModePrivateBrowsing = policy
-
-        verify(mockRuntime.settings.contentBlocking, never()).setCookieBannerModePrivateBrowsing(policy.mode)
-    }
-
-    @Test
-    fun `setCookieBannerHandlingDetectOnlyMode is only invoked when the value is changed`() {
-        val mockRuntime = mock<GeckoRuntime>()
-        val settings = spy(ContentBlocking.Settings.Builder().build())
-        whenever(mockRuntime.settings).thenReturn(mock())
-        whenever(mockRuntime.settings.contentBlocking).thenReturn(settings)
-
-        val engine = GeckoEngine(testContext, runtime = mockRuntime)
-
-        engine.settings.cookieBannerHandlingDetectOnlyMode = true
-
-        verify(mockRuntime.settings.contentBlocking).setCookieBannerDetectOnlyMode(true)
-
-        reset(settings)
-
-        engine.settings.cookieBannerHandlingDetectOnlyMode = true
-
-        verify(mockRuntime.settings.contentBlocking, never()).setCookieBannerDetectOnlyMode(true)
-    }
-
-    @Test
-    fun `setCookieBannerHandlingGlobalRules is only invoked when the value is changed`() {
-        val mockRuntime = mock<GeckoRuntime>()
-        val settings = spy(ContentBlocking.Settings.Builder().build())
-        whenever(mockRuntime.settings).thenReturn(mock())
-        whenever(mockRuntime.settings.contentBlocking).thenReturn(settings)
-
-        val engine = GeckoEngine(testContext, runtime = mockRuntime)
-
-        engine.settings.cookieBannerHandlingGlobalRules = true
-
-        verify(mockRuntime.settings.contentBlocking).setCookieBannerGlobalRulesEnabled(true)
-
-        reset(settings)
-
-        engine.settings.cookieBannerHandlingGlobalRules = true
-
-        verify(mockRuntime.settings.contentBlocking, never()).setCookieBannerGlobalRulesEnabled(true)
-    }
-
-    @Test
-    fun `setCookieBannerHandlingGlobalRulesSubFrames is only invoked when the value is changed`() {
-        val mockRuntime = mock<GeckoRuntime>()
-        val settings = spy(ContentBlocking.Settings.Builder().build())
-        whenever(mockRuntime.settings).thenReturn(mock())
-        whenever(mockRuntime.settings.contentBlocking).thenReturn(settings)
-
-        val engine = GeckoEngine(testContext, runtime = mockRuntime)
-
-        engine.settings.cookieBannerHandlingGlobalRulesSubFrames = true
-
-        verify(mockRuntime.settings.contentBlocking).setCookieBannerGlobalRulesSubFramesEnabled(true)
-
-        reset(settings)
-
-        engine.settings.cookieBannerHandlingGlobalRulesSubFrames = true
-
-        verify(mockRuntime.settings.contentBlocking, never()).setCookieBannerGlobalRulesSubFramesEnabled(true)
-    }
-
-    @Test
     fun `setQueryParameterStripping is only invoked when the value is changed`() {
         val mockRuntime = mock<GeckoRuntime>()
         val settings = spy(ContentBlocking.Settings.Builder().build())
@@ -872,13 +765,6 @@ class GeckoEngineTest {
         engine.settings.emailTrackerBlockingPrivateBrowsing = true
 
         verify(mockRuntime.settings.contentBlocking, never()).setEmailTrackerBlockingPrivateBrowsing(true)
-    }
-
-    @Test
-    fun `Cookie banner handling settings are aligned`() {
-        assertEquals(ContentBlocking.CookieBannerMode.COOKIE_BANNER_MODE_DISABLED, EngineSession.CookieBannerHandlingMode.DISABLED.mode)
-        assertEquals(ContentBlocking.CookieBannerMode.COOKIE_BANNER_MODE_REJECT, EngineSession.CookieBannerHandlingMode.REJECT_ALL.mode)
-        assertEquals(ContentBlocking.CookieBannerMode.COOKIE_BANNER_MODE_REJECT_OR_ACCEPT, EngineSession.CookieBannerHandlingMode.REJECT_OR_ACCEPT_ALL.mode)
     }
 
     @Test
@@ -1074,9 +960,6 @@ class GeckoEngineTest {
         engine.settings.trackingProtectionPolicy = TrackingProtectionPolicy.none()
 
         assertEquals(CookiePolicy.ACCEPT_ALL.id, contentBlockingSettings.cookieBehavior)
-
-        assertEquals(EngineSession.CookieBannerHandlingMode.DISABLED.mode, contentBlockingSettings.cookieBannerMode)
-        assertEquals(EngineSession.CookieBannerHandlingMode.DISABLED.mode, contentBlockingSettings.cookieBannerModePrivateBrowsing)
     }
 
     @Test