+++ components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/ext/TrackingProtectionPolicyKtTest.kt
@@ -23,8 +23,6 @@ class TrackingProtectionPolicyKtTest {
     fun `transform the policy to a GeckoView ContentBlockingSetting`() {
         val policy = TrackingProtectionPolicy.recommended()
         val setting = policy.toContentBlockingSetting()
-        val cookieBannerSetting = EngineSession.CookieBannerHandlingMode.REJECT_OR_ACCEPT_ALL
-        val cookieBannerSettingPrivateBrowsing = EngineSession.CookieBannerHandlingMode.DISABLED
 
         assertEquals(policy.getEtpLevel(), setting.enhancedTrackingProtectionLevel)
         assertEquals(policy.getAntiTrackingPolicy(), setting.antiTrackingCategories)
@@ -33,9 +31,6 @@ class TrackingProtectionPolicyKtTest {
         assertEquals(defaultSafeBrowsing.sumOf { it.id }, setting.safeBrowsingCategories)
         assertEquals(setting.strictSocialTrackingProtection, policy.strictSocialTrackingProtection)
         assertEquals(setting.cookiePurging, policy.cookiePurging)
-        assertEquals(EngineSession.CookieBannerHandlingMode.DISABLED.mode, setting.cookieBannerMode)
-        assertEquals(EngineSession.CookieBannerHandlingMode.REJECT_ALL.mode, setting.cookieBannerModePrivateBrowsing)
-        assertFalse(setting.cookieBannerDetectOnlyMode)
         assertFalse(setting.queryParameterStrippingEnabled)
         assertFalse(setting.queryParameterStrippingPrivateBrowsingEnabled)
         assertEquals("", setting.queryParameterStrippingAllowList[0])
@@ -44,22 +39,12 @@ class TrackingProtectionPolicyKtTest {
         val policyWithSafeBrowsing =
             TrackingProtectionPolicy.recommended().toContentBlockingSetting(
                 safeBrowsingPolicy = emptyArray(),
-                cookieBannerHandlingMode = cookieBannerSetting,
-                cookieBannerHandlingModePrivateBrowsing = cookieBannerSettingPrivateBrowsing,
-                cookieBannerHandlingDetectOnlyMode = true,
-                cookieBannerGlobalRulesEnabled = true,
-                cookieBannerGlobalRulesSubFramesEnabled = true,
                 queryParameterStripping = true,
                 queryParameterStrippingPrivateBrowsing = true,
                 queryParameterStrippingAllowList = "AllowList",
                 queryParameterStrippingStripList = "StripList",
             )
         assertEquals(0, policyWithSafeBrowsing.safeBrowsingCategories)
-        assertEquals(cookieBannerSetting.mode, policyWithSafeBrowsing.cookieBannerMode)
-        assertEquals(cookieBannerSettingPrivateBrowsing.mode, policyWithSafeBrowsing.cookieBannerModePrivateBrowsing)
-        assertTrue(policyWithSafeBrowsing.cookieBannerDetectOnlyMode)
-        assertTrue(policyWithSafeBrowsing.cookieBannerGlobalRulesEnabled)
-        assertTrue(policyWithSafeBrowsing.cookieBannerGlobalRulesSubFramesEnabled)
         assertTrue(policyWithSafeBrowsing.queryParameterStrippingEnabled)
         assertTrue(policyWithSafeBrowsing.queryParameterStrippingPrivateBrowsingEnabled)
         assertEquals("AllowList", policyWithSafeBrowsing.queryParameterStrippingAllowList[0])