+++ components/browser/state/src/main/java/mozilla/components/browser/state/state/TabSessionState.kt
@@ -5,7 +5,6 @@
 package mozilla.components.browser.state.state
 
 import mozilla.components.concept.engine.EngineSession
-import mozilla.components.concept.engine.EngineSession.CookieBannerHandlingStatus
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.storage.HistoryMetadataKey
@@ -18,7 +17,6 @@ import java.util.UUID
  * @property content the [ContentState] of this tab.
  * @property trackingProtection the [TrackingProtectionState] of this tab.
  * @property translationsState the [TranslationsState] of this tab.
- * @property cookieBanner the [CookieBannerHandlingStatus] of this tab.
  * @property parentId the parent ID of this tab or null if this tab has no
  * parent. The parent tab is usually the tab that initiated opening this
  * tab (e.g. the user clicked a link with target="_blank" or selected
@@ -43,7 +41,6 @@ data class TabSessionState(
     override val content: ContentState,
     override val trackingProtection: TrackingProtectionState = TrackingProtectionState(),
     override val translationsState: TranslationsState = TranslationsState(),
-    override val cookieBanner: CookieBannerHandlingStatus = CookieBannerHandlingStatus.NO_DETECTED,
     override val engineState: EngineState = EngineState(),
     override val extensionState: Map<String, WebExtensionState> = emptyMap(),
     override val mediaSessionState: MediaSessionState? = null,
@@ -69,7 +66,6 @@ data class TabSessionState(
         extensionState: Map<String, WebExtensionState>,
         mediaSessionState: MediaSessionState?,
         contextId: String?,
-        cookieBanner: CookieBannerHandlingStatus,
     ): SessionState = copy(
         id = id,
         content = content,
@@ -79,7 +75,6 @@ data class TabSessionState(
         extensionState = extensionState,
         mediaSessionState = mediaSessionState,
         contextId = contextId,
-        cookieBanner = cookieBanner,
     )
 }