+++ components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -79,11 +79,6 @@ abstract class EngineSession(
         ) = Unit
 
         fun onTrackerBlockingEnabledChange(enabled: Boolean) = Unit
-
-        /**
-         * Event to indicate a new [CookieBannerHandlingStatus] is available.
-         */
-        fun onCookieBannerChange(status: CookieBannerHandlingStatus) = Unit
         fun onTrackerBlocked(tracker: Tracker) = Unit
         fun onTrackerLoaded(tracker: Tracker) = Unit
         fun onNavigateBack() = Unit
@@ -684,47 +679,6 @@ abstract class EngineSession(
     }
 
     /**
-     * Represents settings options for cookie banner handling.
-     */
-    @Suppress("MagicNumber")
-    enum class CookieBannerHandlingMode(val mode: Int) {
-        /**
-         * The feature is turned off and cookie banners are not handled
-         */
-        DISABLED(0),
-
-        /**
-         * Reject cookies if possible
-         */
-        REJECT_ALL(1),
-
-        /**
-         * Reject cookies if possible. If rejecting is not possible, accept cookies
-         */
-        REJECT_OR_ACCEPT_ALL(2),
-    }
-
-    /**
-     * Represents a status for cookie banner handling.
-     */
-    enum class CookieBannerHandlingStatus {
-        /**
-         * Indicates a cookie banner was detected.
-         */
-        DETECTED,
-
-        /**
-         * Indicates a cookie banner was handled.
-         */
-        HANDLED,
-
-        /**
-         * Indicates a cookie banner has not been detected yet.
-         */
-        NO_DETECTED,
-    }
-
-    /**
      * Represents settings options for bounce tracking protection.
      */
     enum class BounceTrackingProtectionMode(val mode: Int) {
@@ -1005,16 +959,6 @@ abstract class EngineSession(
      * Enables/disables Desktop Mode with an optional ability to reload the session right after.
      */
     abstract fun toggleDesktopMode(enable: Boolean, reload: Boolean = false)
-
-    /**
-     * Checks if there is a rule for handling a cookie banner for the current website in the session.
-     *
-     * @param onSuccess callback invoked if the engine API returned a valid response. Please note
-     * that the response can be null - which can indicate a bug, a miscommunication
-     * or other unexpected failure.
-     * @param onError callback invoked if there was an error getting the response.
-     */
-    abstract fun hasCookieBannerRuleForSession(onResult: (Boolean) -> Unit, onException: (Throwable) -> Unit)
 
     /**
      * Checks if the current session is using a PDF viewer.