+++ components/concept/engine/src/main/java/mozilla/components/concept/engine/Settings.kt
@@ -4,7 +4,6 @@
 
 package mozilla.components.concept.engine
 
-import mozilla.components.concept.engine.EngineSession.CookieBannerHandlingMode
 import mozilla.components.concept.engine.EngineSession.SafeBrowsingPolicy
 import mozilla.components.concept.engine.EngineSession.TrackingProtectionPolicy
 import mozilla.components.concept.engine.fission.WebContentIsolationStrategy
@@ -52,34 +51,9 @@ abstract class Settings {
     open var trackingProtectionPolicy: TrackingProtectionPolicy? by UnsupportedSetting()
 
     /**
-     * Setting to control the cookie banner handling feature.
-     */
-    open var cookieBannerHandlingMode: CookieBannerHandlingMode by UnsupportedSetting()
-
-    /**
-     * Setting to control the cookie banner handling feature in the private browsing mode.
-     */
-    open var cookieBannerHandlingModePrivateBrowsing: CookieBannerHandlingMode by UnsupportedSetting()
-
-    /**
      * Setting to control tracking protection.
      */
     open var safeBrowsingPolicy: Array<SafeBrowsingPolicy> by UnsupportedSetting()
-
-    /**
-     * Setting to control the cookie banner handling feature detect only mode.
-     */
-    open var cookieBannerHandlingDetectOnlyMode: Boolean by UnsupportedSetting()
-
-    /**
-     * Setting to control the cookie banner handling global rules feature.
-     */
-    open var cookieBannerHandlingGlobalRules: Boolean by UnsupportedSetting()
-
-    /**
-     * Setting to control the cookie banner handling global rules subFrames feature.
-     */
-    open var cookieBannerHandlingGlobalRulesSubFrames: Boolean by UnsupportedSetting()
 
     /**
      * Setting to control the cookie banner enables / disables the URL query string
@@ -497,12 +471,6 @@ data class DefaultSettings(
     override var fdlibmMathEnabled: Boolean = false,
     override var baselineFingerprintingProtection: Boolean? = null,
     override var baselineFingerprintingProtectionOverrides: String? = null,
-    override var cookieBannerHandlingMode: CookieBannerHandlingMode = CookieBannerHandlingMode.DISABLED,
-    override var cookieBannerHandlingModePrivateBrowsing: CookieBannerHandlingMode =
-        CookieBannerHandlingMode.DISABLED,
-    override var cookieBannerHandlingDetectOnlyMode: Boolean = false,
-    override var cookieBannerHandlingGlobalRules: Boolean = false,
-    override var cookieBannerHandlingGlobalRulesSubFrames: Boolean = false,
     override var queryParameterStripping: Boolean = false,
     override var queryParameterStrippingPrivateBrowsing: Boolean = false,
     override var queryParameterStrippingAllowList: String = "",