+++ components/concept/engine/src/test/java/mozilla/components/concept/engine/EngineSessionTest.kt
@@ -1006,11 +1006,6 @@ open class DummyEngineSession : EngineSession() {
 
     override fun toggleDesktopMode(enable: Boolean, reload: Boolean) {}
 
-    override fun hasCookieBannerRuleForSession(
-        onResult: (Boolean) -> Unit,
-        onException: (Throwable) -> Unit,
-    ) {}
-
     override fun checkForPdfViewer(
         onResult: (Boolean) -> Unit,
         onException: (Throwable) -> Unit,